Fix for failure to send events on subsequent non-list combo entries.
The culprit was a bad equality test for the new value vs old value in `_updateAndDispatch()`.

When multiple non-list combo box entries are set in sequence, the new `__value` will be the same as the old `__value` (a value of `undefined`) even though the `__label` changed. The old equality test treated them as equal and incorrectly skipped the update and notification logic.