Fix to ensure re-filtering of the list when input doesn't match an item.
When opening the dropdown, filters the options if the input doesn't match the currently selected value's label (and its not a non-list value).

Think this maintaines the delicate balance here to retain the behavior of showing the whole list if an item has been selected (vs a partial value being typed and the list being filtered).