Switch from down + click events to tap event for iron-dropdown items.
Otherwise touch devices (phones and tablets) are unable to scroll the dropdown. Polymer will synthesize a tap event for mouse environments.

Also stopped event propagation so the "tap" doesn't register through to other controls underneath the dropdown.  For example, prior to this fix, tapping an item in a `radium-combo` used in a `radium-filter-panel` would not only select the item but would also incorrectly trigger collapse or focus on the field that appeared under it.